Fix movecoords bug and wrap listeners in "try".
The comment that used to be in the touchstart listener about double taps
was bogus.  I was confused because I didn't realize that the touchend
listener was erroring out when it tried to index movecoords (undefined
since there had been no intervening touchmove).

Now when a listener errors out it's obvious (it puts up an alert box).